Added robots txt file and refactoring categories entities
