+++ src/Entity/Category.php
@@ -2,10 +2,10 @@
 
 namespace App\Entity;
 
+use App\Entity\Traits\CategoryTrait;
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
 use Doctrine\ORM\Mapping as ORM;
-use Symfony\Component\Validator\Constraints as Assert;
 
 
 /**
@@ -13,50 +13,14 @@ use Symfony\Component\Validator\Constraints as Assert;
  */
 class Category
 {
+    use CategoryTrait;
+
     /**
      * @ORM\Id()
      * @ORM\GeneratedValue()
      * @ORM\Column(type="integer")
      */
     private $id;
-
-    /**
-     * @Assert\NotBlank
-     * @Assert\Length(
-     *      min = 2,
-     *      max = 64,
-     *      minMessage = "Votre nom de catégorie doit comporter {{ limit }} caractères minimum",
-     *      maxMessage = "Votre nom de catégorie doit comporter {{ limit }} caractères maximum"
-     * )
-     *  
-     * @ORM\Column(type="string", length=64)
-     */
-    private $name;
-
-    /**
-     * @ORM\Column(type="string", length=255, nullable=true)
-     */
-    private $image;
-
-    /**
-     * @ORM\Column(type="boolean")
-     */
-    private $enable;
-
-    /**
-     * @Assert\NotBlank
-     * @Assert\Type("\DateTime")
-     * 
-     * @ORM\Column(type="datetime")
-     */
-    private $createdAt;
-
-    /**
-     * @Assert\Type("\DateTime")
-     * 
-     * @ORM\Column(type="datetime", nullable=true)
-     */
-    private $updatedAt;
 
     /**
      * @ORM\OneToMany(targetEntity="App\Entity\Subcategory", mappedBy="category", orphanRemoval=true)
@@ -77,74 +41,9 @@ class Category
         $this->subcategories = new ArrayCollection();
     }
 
-    public function __toString()
-    {
-        return $this->name;
-    }
-
     public function getId(): ?int
     {
         return $this->id;
-    }
-
-    public function getName(): ?string
-    {
-        return $this->name;
-    }
-
-    public function setName(string $name): self
-    {
-        $this->name = $name;
-
-        return $this;
-    }
-
-    public function getImage(): ?string
-    {
-        return $this->image;
-    }
-
-    public function setImage(?string $image): self
-    {
-        $this->image = $image;
-
-        return $this;
-    }
-
-    public function getEnable(): ?bool
-    {
-        return $this->enable;
-    }
-
-    public function setEnable(bool $enable): self
-    {
-        $this->enable = $enable;
-
-        return $this;
-    }
-
-    public function getCreatedAt(): ?\DateTimeInterface
-    {
-        return $this->createdAt;
-    }
-
-    public function setCreatedAt(\DateTimeInterface $createdAt): self
-    {
-        $this->createdAt = $createdAt;
-
-        return $this;
-    }
-
-    public function getUpdatedAt(): ?\DateTimeInterface
-    {
-        return $this->updatedAt;
-    }
-
-    public function setUpdatedAt(?\DateTimeInterface $updatedAt): self
-    {
-        $this->updatedAt = $updatedAt;
-
-        return $this;
     }
 
     /**
